Account

+++ app/Services/Products/BrandService.php
@@ -11,27 +11,44 @@ class BrandService
 {
     public function brandsTable()
     {
-        $brands = DB::table('brands')->orderBy('id', 'desc')->get();
-        $imgUrl = asset('uploads/brand/');
+        $brands = DB::table('brands')->orderBy('id', 'desc');
 
         return DataTables::of($brands)
             ->addIndexColumn()
-            ->editColumn('photo', function ($row) use ($imgUrl) {
-                return '<img loading="lazy" class="rounded img-thumbnail" style="height:30px; width:30px;" src="' . $imgUrl . '/' . $row->photo . '">';
+            ->editColumn('photo', function ($row) {
+
+                $photo = asset('images/general_default.png');
+
+                if ($row->photo) {
+
+                    $photo = asset('uploads/brand/' . $row->photo);
+                }
+                return '<img loading="lazy" class="rounded img-thumbnail" style="height:30px; width:30px;"  src="' . $photo . '">';
             })
             ->addColumn('action', function ($row) {
+
                 $html = '<div class="dropdown table-dropdown">';
-                $html .= '<a href="' . route('brands.edit', [$row->id]) . '" class="action-btn c-edit" id="edit" title="Edit"><span class="fas fa-edit"></span></a>';
-                $html .= '<a href="' . route('brands.delete', [$row->id]) . '" class="action-btn c-delete" id="delete" title="Delete"><span class="fas fa-trash "></span></a>';
+
+                if (auth()->user()->can('product_brand_edit')) {
+
+                    $html .= '<a href="' . route('brands.edit', [$row->id]) . '" class="action-btn c-edit" id="edit" title="Edit"><span class="fas fa-edit"></span></a>';
+                }
+
+                if (auth()->user()->can('product_brand_delete')) {
+
+                    $html .= '<a href="' . route('brands.delete', [$row->id]) . '" class="action-btn c-delete" id="delete" title="Delete"><span class="fas fa-trash "></span></a>';
+                }
                 $html .= '</div>';
 
                 return $html;
             })->rawColumns(['photo', 'action'])->make(true);
     }
 
-    public function addBrand(object $request): ?object
+    public function addBrand(object $request, object $codeGenerator): ?object
     {
+        $code = $codeGenerator->brandCode();
         $addBrand = new Brand();
+        $addBrand->code = $code;
         $addBrand->name = $request->name;
 
         if ($request->file('photo')) {
@@ -54,12 +71,9 @@ class BrandService
 
         if ($request->file('photo')) {
 
-            if ($updateBrand->photo !== 'default.png') {
+            if (isset($updateBrand->photo) && file_exists(public_path('uploads/brand/' . $updateBrand->photo))) {
 
-                if (file_exists(public_path('uploads/brand/' . $updateBrand->photo))) {
-
-                    unlink(public_path('uploads/brand/' . $updateBrand->photo));
-                }
+                unlink(public_path('uploads/brand/' . $updateBrand->photo));
             }
 
             $brandPhoto = $request->file('photo');
@@ -79,12 +93,9 @@ class BrandService
 
         if (isset($deleteBrand)) {
 
-            if ($deleteBrand->photo !== 'default.png') {
+            if (isset($deleteBrand->photo) && file_exists(public_path('uploads/brand/' . $deleteBrand->photo))) {
 
-                if (file_exists(public_path('uploads/brand/' . $deleteBrand->photo))) {
-
-                    unlink(public_path('uploads/brand/' . $deleteBrand->photo));
-                }
+                unlink(public_path('uploads/brand/' . $deleteBrand->photo));
             }
 
             $deleteBrand->delete();
